fix: export all files from sd-jwt-vc package
